Adding the new tests in part3 file
